% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bcgw_connect}
\alias{bcgw_connect}
\title{Connect to the BCGW (LRDW)}
\usage{
bcgw_connect(user, password, ...)
}
\arguments{
\item{user}{BCGW Username}

\item{password}{BCGW password}

\item{...}{Additional parameters passed on to odbcConnect}
}
\value{
an odbc connection
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

