% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_print_ver}
\alias{create_print_ver}
\title{Create a print version starter based on the template}
\usage{
create_print_ver(bucket, title, path)
}
\arguments{
\item{bucket}{Indicator topic (one of Air, Climate Change, Contaminants,
Forests, Land, Plants and Animals, Sustainability, Waste, Water)}

\item{title}{Title of the indicator}

\item{path}{path and filename of the resulting .Rmd file}
}
\description{
<full description of function>
}
\examples{
\dontrun{

}
}

