% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{draft_print_ver}
\alias{draft_print_ver}
\title{Create a draft print version based on the template}
\usage{
draft_print_ver(filename)
}
\arguments{
\item{filename}{the name of the Rmd file you wish to create}
}
\value{
The filename of the new document (invisibly)
}
\description{
Create a draft print version based on the template
}
\examples{
\dontrun{
draft_print_ver("print_ver.Rmd")
}
}
