% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envreportbc-package.r
\docType{package}
\name{envreportbc}
\alias{envreportbc}
\alias{envreportbc-package}
\title{envreportbc}
\description{
The envreportbc package contains templates & internal utility functions for Environmental Reporting BC.
}
