% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envreportbc-package.r
\docType{package}
\name{envreportutils.internal}
\alias{envreportutils.internal}
\alias{envreportutils.internal-package}
\title{envreportutils.internal}
\description{
The envreportutils.internal package contains templates & internal utility functions for Environmental Reporting BC.
}
