% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{html_md}
\alias{html_md}
\title{Convert html text to Rmd for print version}
\usage{
html_md(html, md = tempfile(fileext = "md"))
}
\arguments{
\item{html}{path to the html file to convert from}

\item{rmd}{path to target rmd (should not exist yet)}
}
\value{
an Rmd document
}
\description{
Convert html text to Rmd for print version
}

