% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_local.R
\name{install_dev}
\alias{install_dev}
\title{Install a local envreportbc package}
\usage{
install_dev(pkgname, path = "D:/packages")
}
\arguments{
\item{pkgname}{The name of the package}

\item{path}{= NULL the path where the package is stored. Defaults to default location on the I drive.}
}
\value{
None
}
\description{
<full description>
}
\examples{
\dontrun{
install_dev("envreportutils")
}
}

