% Generated by roxygen2 (4.0.2): do not edit by hand
\name{local_packages}
\alias{local_packages}
\title{List available packages developed by Environmental Reporting BC}
\usage{
local_packages(path = NULL)
}
\arguments{
\item{path}{= NULL The path where packages are stored. Defaults to default location on the I drive.}
}
\value{
a data frame of packages, versions, and locations
}
\description{
List available packages developed by Environmental Reporting BC
}
\examples{
\dontrun{
 local_packages()
}
}

