% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{print_ver}
\alias{print_ver}
\title{PDF specifications for EnvReportBC print version}
\usage{
print_ver(..., keep_tex = FALSE)
}
\arguments{
\item{...}{arguments passed on to pdf_document}

\item{keep_tex}{keep tex file?}
}
\value{
pdf_document specification
}
\description{
PDF specifications for EnvReportBC print version
}

