% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_template.R
\name{slide_template}
\alias{slide_template}
\title{Create a slide presentation from template}
\usage{
slide_template(title, path = ".")
}
\arguments{
\item{title}{Short title of the presentation}

\item{path}{where do you want to put it? (default current directory)}
}
\value{
TRUE (invisibly)
}
\description{
Create a slide presentation from template
}
\examples{
\dontrun{
  slide_template("my_pres", path = "presentation")
}
}
