% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{web2print_ver}
\alias{web2print_ver}
\title{Convert a file on the SoE wwwd site to markdown in the envreportbc 
rmarkdown template, and save it in the \code{print_ver} folder.}
\usage{
web2print_ver(html_file)
}
\arguments{
\item{html_file}{the file in the soe/indicators folder on the web. Use the form
\code{folder/indicator.html} (e.g., \code{"land/roads.html"})}
}
\value{
writes the file to the \code{print_ver} folder of the current project.
}
\description{
If there is an existing file in the \code{print_ver} folder, a backup will be 
created
}
