% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bcgw_connect.R
\name{bcgw_connect}
\alias{bcgw_connect}
\title{Connect to the BCGW (LRDW)}
\usage{
bcgw_connect(user, ...)
}
\arguments{
\item{user}{BCGW Username}

\item{...}{Additional parameters passed on to odbcConnect}
}
\value{
an odbc connection
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

