% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{create_print_ver}
\alias{create_print_ver}
\title{Create a print version starter based on the template}
\usage{
create_print_ver(html = NULL, name = NULL, dir = "print_ver")
}
\arguments{
\item{html}{Path to the html file to convert from}

\item{name}{Short name which will form the base name of the .Rmd file}

\item{dir}{Folder in which to place resulting .Rmd file. Default is
/code{print_ver} inside the current working directory}
}
\description{
<full description of function>
}

