% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{create_print_ver}
\alias{create_print_ver}
\title{Create a print version starter based on the template}
\usage{
create_print_ver(bucket = NULL, name = NULL, dir = "print_ver")
}
\arguments{
\item{bucket}{Indicator topic (one of Air, Climate Change, Contaminants,
Forests, Land, Plants and Animals, Sustainability, Waste, Water)}

\item{name}{Short name which will form the base name of the .Rmd file}

\item{dir}{Folder in which to place resulting .Rmd file. Default is
/code{print_ver} inside the current working directory}
}
\description{
<full description of function>
}
\examples{
\donttest{

}
}

