% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_print_ver.R
\name{html_rmd}
\alias{html_rmd}
\title{Convert html text to Rmd for print version}
\usage{
html_rmd(html, rmd)
}
\arguments{
\item{html}{}

\item{rmd}{}
}
\value{
an Rmd document
}
\description{
Convert html text to Rmd for print version
}

