% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indicator_skeleton}
\alias{indicator_skeleton}
\title{Creates the framework of a new indicator development folder}
\usage{
indicator_skeleton(path, print_ver = TRUE, bucket, title, rstudio = FALSE)
}
\arguments{
\item{path}{location to create new indicator}

\item{print_ver}{create a print version template?}

\item{bucket}{(required if print_ver = \code{TRUE}) Indicator topic (one of
Air, Climate Change, Contaminants, Forests, Land, Plants and Animals,
Sustainability, Waste, Water)}

\item{title}{(required if print_ver = \code{TRUE}) Title of the indicator}

\item{rstudio}{Create an Rstudio project file?}
}
\description{
Creates the folder structure for a new indicator, including a template for
the printable version RMarkdown document.
}
\examples{
\dontrun{
indicator_skeleton(path = "c:/_dev/tarballs", print_ver = TRUE,
                   bucket="Contaminants",
                   title="Trends in Tar Ball deposition in BC (1876-1921)",
                   rstudio = TRUE)
}
}

