% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indicator_skeleton.R
\name{indicator_skeleton}
\alias{indicator_skeleton}
\title{Creates the framework of a new indicator development folder}
\usage{
indicator_skeleton(path = ".", print_ver = TRUE, bucket, git_init = TRUE,
  git_clone = NULL, rstudio = TRUE)
}
\arguments{
\item{path}{location to create new indicator. If \code{"."} (the default),
the name of the working directory will be taken as the indicator name. If
not \code{"."}, the last component of the given path will be used as the
indicator name.}

\item{print_ver}{create a print version template?}

\item{bucket}{(required if print_ver = \code{TRUE}) Indicator topic (one of
Air, Climate Change, Contaminants, Forests, Land, Plants and Animals,
Sustainability, Waste, Water)}

\item{git_init}{Create a new git repository? Logical, default \code{TRUE}.}

\item{git_clone}{the url of a git repo to clone.}

\item{rstudio}{Create an Rstudio project file?}
}
\description{
Creates the folder structure for a new indicator, including a template for
the printable version RMarkdown document.
}
\details{
If you are cloning a repository (\code{git_clone = "path_to_repo"}),
  you should run this function from the root of your dev folder and leave
  \code{path = "."}, as the repository will be cloned into a new folder. If
  you are setting up a new project (with or without git), \code{path} should
  be \code{"."} if you are within an already created project directory, or
  the name of the folder you want to create.
}
\examples{
\donttest{
 indicator_skeleton(path = "c:/_dev/tarballs", print_ver = TRUE,
                    bucket="Contaminants",
                    rstudio = TRUE)
}
}

