% Generated by roxygen2 (4.0.2): do not edit by hand
\name{install_local}
\alias{install_local}
\title{Install a local envreportbc package}
\usage{
install_local(pkgname, install_path = NULL)
}
\arguments{
\item{pkgname}{The name of the package}

\item{install_path}{= NULL the path where the package is stored. Defaults to default location on the I drive.}
}
\value{
None
}
\description{
<full description>
}
\examples{
\dontrun{
install_local("envreportutils")
}
}

