% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_local.R
\name{local_packages}
\alias{local_packages}
\title{List available packages developed by Environmental Reporting BC}
\usage{
local_packages(path = "D:/packages")
}
\arguments{
\item{path}{The path where packages are stored. Defaults to default location on the I drive.}
}
\value{
a data frame of packages, versions, and locations
}
\description{
List available packages developed by Environmental Reporting BC
}

